% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.data.R
\name{generate.data}
\alias{generate.data}
\title{Generate Peng and Wang (2015) simulation data}
\usage{
generate.data(n, p, signal = 1)
}
\arguments{
\item{n}{number of observations}

\item{p}{number of variables (p > 20)}

\item{signal}{signal of true beta's}
}
\value{
Y simulated n x 1 Y vector

X simulated n x p X matrix

true_beta simulated true beta's
}
\description{
Generate Peng and Wang (2015) simulation data
}
\examples{
set.seed(1)
data = generate.data(n = 150, p = 150, signal = 1)
}
