% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcd.lasso.fit.R
\name{qcd.lasso.fit}
\alias{qcd.lasso.fit}
\title{Solve LASSO penalized quantile regression by coordinate descent for a single value of lambda}
\usage{
qcd.lasso.fit(x, y, tau, lambda, thresh = 1e-06, maxit = 1e+05, warm = NULL)
}
\arguments{
\item{x}{\code{n x p} design matrix X.}

\item{y}{\code{n x 1} vector of response variables Y.}

\item{tau}{Quantile value between 0 and 1.}

\item{lambda}{A single lambda value.}

\item{thresh}{Threshold of checking whether the coefficients converged.}

\item{maxit}{Maximum iteration for convergence.}

\item{warm}{Whether warm start will be used.
Do not use this argument when only running \code{qcd.lasso.fit}.
It is used when constructing regularization path with \code{qcd.path} later.}
}
\value{
\item{beta}{A \code{n x 1} matrix of coefficients.}

\item{dim}{Dimension of coefficient vector.}

\item{lambda}{Lambda value used.}

\item{df}{Number of nonzero coefficients.}
}
\description{
Solve LASSO penalized quantile regression by coordinate descent for a single value of lambda
}
\examples{
## Create sample data set
n <- 30; p <- 30
x <- array(rnorm(n*p), c(n,p))
for (j in 1:p){x[,j] = x[,j]/(norm(as.matrix(x[,j]), type="f")/sqrt(n))}
e <- rnorm(n)
b <- c(1, -1, rep(0, p-2))
y <- x \%*\% b + e

## Use QCD algorithm to solve LASSO penalized quantile regression with one lambda
qr.lasso <- qcd.lasso.fit(x = x, y = y, tau = 0.5, lambda = 0.1,
thresh = 1e-06, maxit = 100)


}
