% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Compute RMSE with estimated beta}
\usage{
rmse(beta, truebeta, lambda)
}
\arguments{
\item{beta}{length(lambda) x p estimated beta matrix}

\item{truebeta}{true beta vector}

\item{lambda}{grid of lambda's}
}
\value{
RMSE value for each lambda
}
\description{
Compute RMSE with estimated beta
}
\examples{
upper = 2; lower = -6
lambda =  2^seq(upper, lower, by = -0.2)
p = 50; len = length(lambda)
estimated_beta = matrix(rnorm(p*len, 0, 1), ncol = p)
true_beta = rnorm(50, 0, 1)
lasso.warm.rmse = rmse(estimated_beta, true_beta, lambda)
}
