% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.data.R
\name{generate.data}
\alias{generate.data}
\title{Generate Peng and Wang (2015) simulation data}
\usage{
generate.data(n, p, signal = 1, tau = 0.5)
}
\arguments{
\item{n}{number of observations.}

\item{p}{number of variables. (p > 20)}

\item{signal}{signal of true beta's.}

\item{tau}{quantile value between 0 and 1.}
}
\value{
\item{Y}{n x 1 vector.}

\item{X}{n x p matrix.}

\item{true_beta}{p x 1 true beta vector.}
}
\description{
Generate Peng and Wang (2015) simulation data
}
\examples{
set.seed(1)
data = generate.data(n = 150, p = 150, signal = 1, tau = 0.5)

}
